Move statics
